+++ day44/models/user.js
@@ -3,20 +3,35 @@ module.exports = class username extends Sequelize.Model {
   static init(sequelize) {
     return super.init(
       {
-        userId: {
+        user_id: {
           type: Sequelize.INTEGER,
           // 객체 형식으로 저장
           allowNull: false,
           autoIncrement: true,
           primaryKey: true,
         },
-        username: {
+        user_pw: {
+          type: Sequelize.STRING,
+          allowNull: false,
+        },
+        user_name: {
           type: Sequelize.STRING(15),
           unique: true,
           allowNull: false,
         },
-        userPassword: {
-          type: Sequelize.STRING,
+        user_number: {
+          type: Sequelize.STRING(15),
+          unique: true,
+          allowNull: false,
+        },
+        user_local: {
+          type: Sequelize.STRING(15),
+          unique: true,
+          allowNull: false,
+        },
+        user_temperature: {
+          type: Sequelize.STRING(15),
+          unique: true,
           allowNull: false,
         },
       },
