+++ day44/public/font/font.css
@@ -1,0 +1,11 @@
+@font-face {
+  font-family: "maple-bold";
+  src: url("./Maplestory\ Bold.ttf") format("truetype");
+  font-weight: bold;
+}
+
+@font-face {
+  font-family: "maple-light";
+  src: url("./Maplestory\ Light.ttf") format("truetype");
+  font-weight: normal;
+}
