--- day44/public/regist/index.html
@@ -1,138 +1,0 @@
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <link rel="stylesheet" href="index.css" />
-    <title>회원가입 페이지</title>
-  </head>
-  <body>
-    <!-- Choose Form -->
-
-    <section class="form-container">
-      <span id="arrowClick" class="form-container__arrow"
-        ><i class="fa fa-arrow-circle-left" aria-hidden="true"></i
-      ></span>
-      <div class="overlay"></div>
-      <div class="choose-form">
-        <div class="title">
-          <h1 class="title__h1">3DAYMarket</h1>
-        </div>
-        <div class="buttons">
-          <a id="signUp" class="buttons__signup" href="#">Sign up</a>
-          <a id="login" class="buttons__signup buttons__signup--login" href="#"
-            >Login</a
-          >
-        </div>
-      </div>
-
-      <!-- Login Form -->
-
-      <div class="login-form">
-        <div class="form-wrapper">
-          <form method="post">
-            <label class="form-wrapper__label" for="login-username"
-              >Username</label
-            >
-            <input
-              id="login-username"
-              class="form-wrapper__input"
-              type="text"
-              placeholder="Username"
-              name="username"
-              required
-            />
-            <label class="form-wrapper__label" for="login-password"
-              >Password</label
-            >
-            <input
-              id="login-password"
-              class="form-wrapper__input"
-              type="password"
-              placeholder="Password"
-              name="password"
-              pattern=".{3,}"
-              title="Password must contain at least 3 characters"
-              required
-            />
-            <a class="login-form__forgot-password" href="#">Forgot password?</a>
-            <button
-              class="buttons__signup buttons__signup--login-form"
-              type="submit"
-              href="#"
-            >
-              Login
-            </button>
-          </form>
-
-          <!-- Social-Media -->
-
-          <div class="social-media">
-            <h6 class="title__h2">Or connect with</h6>
-            <a class="buttons__signup buttons__signup--social" href="#"
-              ><i class="fab fa-facebook-f" aria-hidden="true"></i
-              >&nbsp;facebook</a
-            >
-            <a class="buttons__signup buttons__signup--social" href="#"
-              ><i class="fab fa-google" aria-hidden="true"></i>&nbsp;google</a
-            >
-          </div>
-        </div>
-      </div>
-
-      <!-- Register Form -->
-
-      <div class="login-form login-form--register">
-        <div class="form-wrapper">
-          <form method="post">
-            <label class="form-wrapper__label" for="signup-username"
-              >Username</label
-            >
-            <input
-              id="signup-username"
-              class="form-wrapper__input"
-              type="text"
-              placeholder="Username"
-              name="username"
-              required
-            />
-            <label class="form-wrapper__label" for="signup-password"
-              >Password</label
-            >
-            <input
-              id="signup-password"
-              class="form-wrapper__input"
-              type="password"
-              placeholder="Password"
-              name="password"
-              pattern=".{3,}"
-              title="Password must contain at least 3 characters"
-              required
-            />
-            <button
-              class="buttons__signup buttons__signup--sign-up-form"
-              type="submit"
-            >
-              Sign up
-            </button>
-          </form>
-
-          <!-- Social-Media -->
-
-          <div class="social-media">
-            <h2 class="title__h2">Or connect with</h2>
-            <a class="buttons__signup buttons__signup--social" href="#"
-              ><i class="fab fa-facebook-f" aria-hidden="true"></i
-              >&nbsp;facebook</a
-            >
-            <a class="buttons__signup buttons__signup--social" href="#"
-              ><i class="fab fa-google" aria-hidden="true"></i>&nbsp;google</a
-            >
-          </div>
-        </div>
-      </div>
-    </section>
-    <script src="index.js"></script>
-  </body>
-</html>
