+++ day44/routes/index.js
@@ -1,5 +1,5 @@
 const router = require("express").Router();
-// const user = require("./user.js");
+const user = require("./user.js");
 // router.use("/user", user);
 
 module.exports = router;
